test/web-platform-tests/to-upstream/html/semantics/tabular-data/the-th-element/class-hierarchy.html

<!DOCTYPE HTML>
<title>HTMLTableCellElement (for &lt;th&gt;)</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/tables.html#the-th-element">
<link rel="author" title="Timothy Gu" href="mailto:timothygu99@gmail.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
"use strict";
test(() => {
  const th = document.createElement("th");
  assert_true(th instanceof HTMLTableCellElement, "th instanceof HTMLTableCellElement");
}, "<th> is an instance of HTMLTableCellElement");

test(() => {
  const th = document.createElement("th");
  assert_equals(Object.getPrototypeOf(th), HTMLTableCellElement.prototype,
                "th.__proto__ === HTMLTableCellElement.prototype");
}, "<th> is not an instance of a class inheriting HTMLTableCellElement");
</script>
